[CMake] Make oneMath dependecy optional
Currently when trying to build the whole project using CMake without a
oneMath install on the system, CMake will configure okay but then
fail to find dependencies at build time.

A better used experience would be to try detect oneMath at CMake
configure time, and then skip the `oneMath_gemm` target with a
warning if oneMath isn't found.

If we really do want `oneMath` to be a dependency then this should
be `find_package(oneMath REQUIRED)`, but that feels like it increases
the barrier to entry for newcomers.