Use CTest to help with CI testing
The executable targets to run in acpp CI workflow are currently derived from shell
scripting, which is a relatively fragile way to things.

This PR adds a ctest target for each of the executable created by the build, such
that a user can easily run all the binaries by invoking `$ ctest` from the
build directory. ctest was chosen as it has straightforward CMake integration,
and I've tested it locally with both DPC++ and AdaptiveCPP builds of SYCL Academy.

I've updated the GitHub workflow for AdaptiveCpp to use the new functionality,
passing a regex exclude filter for the tests known to be failing.